+++ treex/__init__.py
@@ -7,6 +7,7 @@ from flax.linen import initializers
 from treex.module import *
 from treex.optimizer import *
 from treex.rnq_seq import RngSeq
+from treex.tree_object import *
 from treex.types import *
 
 from . import nn
